Changed the asFUNCTIONPR and asMETHODPR macros to use static_cast instead of C-style cast.
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@414 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/tests/test_feature/source/test_vector3_2.cpp
@@ -57,7 +57,7 @@ inline csVector3 operator* (float f, const csVector3& v)
 { return csVector3(v.x*f, v.y*f, v.z*f); }
 
 /// Divide a vector by a scalar int.
-inline csVector3 operator/ (const csVector3& v, int f)
+inline csVector3 operator/ (const csVector3& v, float f)
 { return v * (1/(float)f); }
 
 void ConstructVector3(csVector3*o)
